(PIE-252) Allow auth to be passed via params
In order to make this module compatible with PE, you can now pass a
username, password, and instance via task parameters rather than an
inventory file. If a username, password, or instance is not provided in
the parameters the values in the inventory file will be used. Parameter
credentials are prioritized over those provided in an inventory file.