Initial attempt at adding external datasource
This change adds the files need to enable this module to act as an
external data source for SNOW data.

It adds a manifest class that will configure puppet to look for a script
that will get the data, and it adds some configuration to allow the
script to run.

At this point we probably still need a better way to set the
configuration and we will need to set about writing the script itself.

All of the previously present SNOW bolt tasks are still in place and
will still provide useful tasks for people doing the integration.