(PIE-501) Create Snowtoken on Init
Refactored the ServiceNow class so it can generate an oauth_token via
providing the client_id and client_secret.
Added a get_token task which uses the init of the ServiceNow class to
return the oauth_token. Added helper functions in the ServiceNow class
and refactored all the tasks to use those helper functions so that the
ServiceNow class takes care of the request details.