(PIE-514) Add task to setup Midserver
Adds install_midserver task that will install the ServiceNow midserver
on a Debian machine. Defaults for testing are supplied for every
parameter except for the instance. These values should be changed if
not being using for a testing environment. Futhermore the servicenow
instance you use will require the same midserver user to be installed
with the same same midserver user password.